REFACTOR -> moving all classes to abstraction package
