+++ src/main/java/SoftEng754_A4/SoftEng754_A4/abstraction/App.java
@@ -1,4 +1,4 @@
-package SoftEng754_A4.SoftEng754_A4;
+package SoftEng754_A4.SoftEng754_A4.abstraction;
 
 /**
  * Hello world!
