+++ src/main/java/SoftEng754_A4/SoftEng754_A4/abstraction/GitReader.java
@@ -1,4 +1,4 @@
-package SoftEng754_A4.SoftEng754_A4;
+package SoftEng754_A4.SoftEng754_A4.abstraction;
 
 import java.util.HashMap;
 import java.util.Map;
