+++ src/test/java/SoftEng754_A4/SoftEng754_A4/abstraction/TestAbstraction.java
@@ -1,4 +1,4 @@
-package SoftEng754_A4.SoftEng754_A4;
+package SoftEng754_A4.SoftEng754_A4.abstraction;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,6 +7,8 @@ import java.util.Set;
 
 import org.mockito.Mockito;
 
+import SoftEng754_A4.SoftEng754_A4.abstraction.Abstraction;
+import SoftEng754_A4.SoftEng754_A4.abstraction.Abstractor;
 import junit.framework.TestCase;
 import scala.collection.mutable.HashSet;
 
@@ -79,13 +81,6 @@ public class TestAbstraction extends TestCase{
 	//	Test getNamesOfCommitedFiles
 	public void testGetNamesOfCommitedFiles() {
 		
-//		Set expectedSet = new ;
-//		expectedSet.add("abc");
-//		expectedSet.add("bcd");
-//		expectedSet.add("cde");
-//		expectedSet.add("def");
-//		
-//		
 		Abstractor abstractor = Mockito.mock(Abstractor.class);
 		TestAbstraction testObject = Mockito.mock(TestAbstraction.class);
 		
@@ -94,6 +89,7 @@ public class TestAbstraction extends TestCase{
 		assertEquals(checkMapKeySet(abstractor.getCommitedFileNames(), new HashSet<String>()), true);
 		
 	}
+	
 	//	Test getCommitAbstract
 	//	Test getFileAbstract
 	//	Test getTempFileFromRepo
